Split up transaction flow
