+++ app/src/main/res/layout/layout_transaction_details_asset_transfer.xml
@@ -13,7 +13,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:padding="16dp"
-        android:text="@string/generic_transaction_description" />
+        android:text="@string/token_transaction_description" />
 
     <View
         android:layout_width="match_parent"
