+++ app/src/main/res/layout/merge_transaction_details_container.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<merge>
+
+    <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        android:id="@+id/merge_transaction_details_container"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content" />
+</merge>
