+++ app/src/main/res/values/strings.xml
@@ -229,4 +229,5 @@
     <string name="restore_account_description">Type your 12 words from the recovery phrase to restore your account.</string>
     <string name="enter_master_password">Enter Master Password</string>
     <string name="mnemonic_required_words">%1d/%2d words</string>
+    <string name="review_transaction">Review transaction</string>
 </resources>
